vpn-example/requirements.txt
-e .
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ecs_patterns
aws-cdk.aws_rds
aws-cdk.core
aws_cdk.aws_secretsmanager
